+++ ui/src/app/core/core.module.ts
@@ -7,14 +7,12 @@ import { HrefTargetBlankDirective } from './directives/href-target-blank.directi
 import { LongClickDirective } from './directives/longclick.directive';
 import { ResetHomebridgeModalComponent } from './reset-homebridge-modal/reset-homebridge-modal.component';
 import { TranslateModule } from '@ngx-translate/core';
-import { ResetCachedAccessoriesModalComponent } from './reset-cached-accessories-modal/reset-cached-accessories-modal.component';
 import { BackupRestoreComponent } from './backup-restore/backup-restore.component';
 import { ConfirmComponent } from './components/confirm/confirm.component';
 
 @NgModule({
   entryComponents: [
     ResetHomebridgeModalComponent,
-    ResetCachedAccessoriesModalComponent,
     BackupRestoreComponent,
     ConfirmComponent,
   ],
@@ -25,7 +23,6 @@ import { ConfirmComponent } from './components/confirm/confirm.component';
     HrefTargetBlankDirective,
     LongClickDirective,
     ResetHomebridgeModalComponent,
-    ResetCachedAccessoriesModalComponent,
     BackupRestoreComponent,
     ConfirmComponent,
   ],