+++ ui/src/app/modules/settings/remove-all-cached-accessories-modal/remove-all-cached-accessories-modal.component.ts
@@ -1,23 +1,22 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { TranslateService } from '@ngx-translate/core';
 import { ToastrService } from 'ngx-toastr';
-import { ApiService } from '../api.service';
+import { ApiService } from '../../../core/api.service';
 
 @Component({
-  selector: 'app-reset-cached-accessories-modal',
-  templateUrl: './reset-cached-accessories-modal.component.html',
-  styleUrls: ['./reset-cached-accessories-modal.component.scss'],
+  selector: 'app-remove-all-cached-accessories-modal',
+  templateUrl: './remove-all-cached-accessories-modal.component.html',
+  styleUrls: ['./remove-all-cached-accessories-modal.component.scss'],
 })
-export class ResetCachedAccessoriesModalComponent {
+export class RemoveAllCachedAccessoriesModalComponent {
   public clicked: boolean;
 
   constructor(
     public activeModal: NgbActiveModal,
     public toastr: ToastrService,
     private translate: TranslateService,
-    private $route: Router,
     private $api: ApiService,
   ) { }
 
@@ -28,7 +27,7 @@ export class ResetCachedAccessoriesModalComponent {
         this.toastr.success(this.translate.instant('reset.toast_clear_cached_accessories_success'), this.translate.instant('toast.title_success'));
         this.activeModal.close();
       },
-      async err => {
+      err => {
         this.toastr.error(this.translate.instant('reset.toast_failed_to_reset'), this.translate.instant('toast.title_error'));
       },
     );