+++ ui/src/app/modules/settings/settings.component.ts
@@ -8,8 +8,9 @@ import { debounceTime } from 'rxjs/operators';
 
 import { AuthService } from '../../core/auth/auth.service';
 import { ApiService } from '../../core/api.service';
-import { ResetCachedAccessoriesModalComponent } from '../../core/reset-cached-accessories-modal/reset-cached-accessories-modal.component';
+import { RemoveAllCachedAccessoriesModalComponent } from './remove-all-cached-accessories-modal/remove-all-cached-accessories-modal.component';
 import { ResetHomebridgeModalComponent } from '../../core/reset-homebridge-modal/reset-homebridge-modal.component';
+import { RemoveSingleCachedAccessoryModalComponent } from './remove-single-cached-accessory-modal/remove-single-cached-accessory-modal.component';
 
 @Component({
   selector: 'app-settings',
@@ -98,8 +99,14 @@ export class SettingsComponent implements OnInit {
     });
   }
 
-  resetCachedAccessories() {
-    this.$modal.open(ResetCachedAccessoriesModalComponent, {
+  removeAllCachedAccessories() {
+    this.$modal.open(RemoveAllCachedAccessoriesModalComponent, {
+      size: 'lg',
+    });
+  }
+
+  removeSingleCachedAccessories() {
+    this.$modal.open(RemoveSingleCachedAccessoryModalComponent, {
       size: 'lg',
     });
   }