+++ ui/src/app/modules/settings/settings.module.ts
@@ -6,9 +6,19 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 import { SettingsRoutingModule } from './settings-routing.module';
 import { SettingsComponent } from './settings.component';
+import { RemoveAllCachedAccessoriesModalComponent } from './remove-all-cached-accessories-modal/remove-all-cached-accessories-modal.component';
+import { RemoveSingleCachedAccessoryModalComponent } from './remove-single-cached-accessory-modal/remove-single-cached-accessory-modal.component';
 
 @NgModule({
-  declarations: [SettingsComponent],
+  entryComponents: [
+    RemoveAllCachedAccessoriesModalComponent,
+    RemoveSingleCachedAccessoryModalComponent,
+  ],
+  declarations: [
+    SettingsComponent,
+    RemoveAllCachedAccessoriesModalComponent,
+    RemoveSingleCachedAccessoryModalComponent,
+  ],
   imports: [
     CommonModule,
     FormsModule,